food_backend/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
bleach==6.1.0
Django==4.2.13
django-cors-headers==4.3.1
django-redis==5.4.0
django-simpleui==2024.4.1
django-summernote==0.8.20.0
django-uuslug==2.0.0
et-xmlfile==1.1.0
mysqlclient==2.2.4
numpy==1.26.4
openpyxl==3.1.3
pandas==2.2.2
pillow==10.3.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
redis==5.0.4
six==1.16.0
sqlparse==0.5.0
text-unidecode==1.3
typing_extensions==4.11.0
tzdata==2024.1
webencodings==0.5.1
